+++ CHANGES.rst
@@ -11,15 +11,17 @@ Version 1.0.0 (in development)
     the value of `host`, which will mock the connection to the server
   - support for Python 3.12
   - type annotations (:PEP:`484` and :PEP:`561` using inline types)
-  - ``freeze32`` console script to create a new 32-bit server
+  - ``freeze32`` console script to create a new server
 
 * Changed
 
   - convert to a :PEP:`420` implicit namespace package
   - the `requires_pythonnet` and `requires_comtypes` arguments to
     :func:`freeze_server32.main() <msl.loadlib.freeze_server32.main>`
-    were removed and the `imports` and `data` arguments were added
-  - constants (e.g., `IS_WINDOWS`) were moved to a `constants.py` file
+    were removed and the `imports`, `data` and `skip_32bit_check`
+    arguments were added
+  - constants (e.g., `IS_WINDOWS`) were moved to a `constants.py` file,
+    these constants are meant for internal use only
 
 * Removed